Arreglados algunos problemas con el equipo y los null

+++ src/main/java/christian/wordle/wordgames/model/Equipo.java
@@ -20,7 +20,7 @@ public class Equipo {
 
     private String nombre;
 
-    private int puntos;
+    private Long puntos;
 
     private String logo = null;
